added header component
