+++ src/components/App.js
@@ -1,6 +1,4 @@
-import {
-  Switch, Route, Redirect, NavLink,
-} from 'react-router-dom';
+import { Switch, Route, Redirect } from 'react-router-dom';
 import CoinList from '../containers/CoinList';
 import Coin from '../containers/Coin';
 import styles from '../styles/App.module.scss';
@@ -9,9 +7,6 @@ import Header from './Header';
 function App() {
   return (
     <div className={styles.App}>
-      <nav>
-        <NavLink to="/">Home</NavLink>
-      </nav>
       <Header />
       <Switch>
         <Route path="/" exact component={CoinList} />
